backend/requirements.txt
Django==4.2.29
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
dj-database-url>=2.1.0
python-dotenv>=1.0.0
